User/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = "Serviço para gerenciamento de usuários e autenticação"
authors = ["Seu Nome <seu.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.116.1,<0.117.0"
uvicorn = {extras = ["standard"], version = ">=0.35.0,<0.36.0"}
gunicorn = "^22.0.0"
pydantic = {extras = ["email"], version = ">=2.11.7,<3.0.0"}
sqlalchemy = ">=2.0.43,<3.0.0"
pydantic-settings = ">=2.10.1,<3.0.0"
alembic = ">=1.16.4,<2.0.0"
pwdlib = {extras = ["argon2"], version = ">=0.2.1,<0.3.0"}
python-multipart = ">=0.0.20,<0.0.21"
pyjwt = ">=2.10.1,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"